fix: numpy type error + remov legacy argument "multichannel"

+++ imagecorruptions/corruptions.py
@@ -39,7 +39,7 @@ def plasma_fractal(mapsize=256, wibbledecay=3):
     'mapsize' must be a power of two.
     """
     assert (mapsize & (mapsize - 1) == 0)
-    maparray = np.empty((mapsize, mapsize), dtype=np.float_)
+    maparray = np.empty((mapsize, mapsize), dtype=np.float32)
     maparray[0, 0] = 0
     stepsize = mapsize
     wibble = 100
